requirements.txt
ffmpeg-python
vosk
tqdm
inquirer
colorama
pygame
